+++ pkg/api/file_repository.go
@@ -16,6 +16,11 @@ type FileSystemRepository struct {
 	sync.RWMutex
 }
 
+// Type used for JSON.Unmarshaller
+type definitionList struct {
+	defs []Definition
+}
+
 // NewFileSystemRepository creates a mongo country repo
 func NewFileSystemRepository(dir string) (*FileSystemRepository, error) {
 	repo := &FileSystemRepository{InMemoryRepository: NewInMemoryRepository()}
@@ -36,9 +41,11 @@ func NewFileSystemRepository(dir string) (*FileSystemRepository, error) {
 			}
 
 			definition := repo.parseDefinition(appConfigBody)
-			if err = repo.Add(definition); err != nil {
-				log.WithError(err).Error("Can't add the definition to the repository")
-				return nil, err
+			for _, v := range definition.defs {
+				if err = repo.Add(&v); err != nil {
+					log.WithField("name", v.Name).WithError(err).Error("Failed during add definition to the repository")
+					return nil, err
+				}
 			}
 		}
 	}
@@ -46,11 +53,21 @@ func NewFileSystemRepository(dir string) (*FileSystemRepository, error) {
 	return repo, nil
 }
 
-func (r *FileSystemRepository) parseDefinition(apiDef []byte) *Definition {
-	appConfig := NewDefinition()
-	if err := json.Unmarshal(apiDef, appConfig); err != nil {
-		log.WithError(err).Error("[RPC] --> Couldn't unmarshal api configuration")
+func (r *FileSystemRepository) parseDefinition(apiDef []byte) definitionList {
+	appConfigs := definitionList{}
+
+	// Try unmarshalling as if json is an unnamed Array of multiple definitions
+	if err := json.Unmarshal(apiDef, &appConfigs); err != nil {
+		// Try unmarshalling as if json is a single Definition
+		appConfigs.defs = append(appConfigs.defs, *NewDefinition())
+		if err := json.Unmarshal(apiDef, &appConfigs.defs[0]); err != nil {
+			log.WithError(err).Error("[RPC] --> Couldn't unmarshal api configuration")
+		}
 	}
 
-	return appConfig
+	return appConfigs
+}
+
+func (d *definitionList) UnmarshalJSON(b []byte) error {
+	return json.Unmarshal(b, &d.defs)
 }