+++ pkg/plugin/oauth2/handlers.go
@@ -76,17 +76,17 @@ func (c *Controller) PutBy() http.HandlerFunc {
 		}
 
 		err = json.NewDecoder(r.Body).Decode(oauth)
-		if nil != err {
+		if err != nil {
 			errors.Handler(w, err)
 			return
 		}
 
-		span = opentracing.FromContext(r.Context(), "datastore.Add")
-		err = c.repo.Add(oauth)
+		span = opentracing.FromContext(r.Context(), "datastore.Save")
+		err = c.repo.Save(oauth)
 		c.dispatch(notifier.NoticeOAuthServerUpdated)
 		span.Finish()
 
-		if nil != err {
+		if err != nil {
 			errors.Handler(w, errors.New(http.StatusBadRequest, err.Error()))
 			return
 		}
@@ -112,7 +112,7 @@ func (c *Controller) Post() http.HandlerFunc {
 		span.Finish()
 
 		if nil != err {
-			errors.Handler(w, errors.New(http.StatusBadRequest, err.Error()))
+			errors.Handler(w, err)
 			return
 		}