+++ pkg/proxy/definition_test.go
@@ -6,12 +6,56 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestNewDefinitions(t *testing.T) {
+func TestDefinition(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		scenario string
+		function func(*testing.T)
+	}{
+		{
+			scenario: "new definitions",
+			function: testNewDefinitions,
+		},
+		{
+			scenario: "successful validation",
+			function: testSuccessfulValidation,
+		},
+		{
+			scenario: "empty listen path validation",
+			function: testEmptyListenPathValidation,
+		},
+		{
+			scenario: "invalid target url validation",
+			function: testInvalidTargetURLValidation,
+		},
+		{
+			scenario: "route to json",
+			function: testRouteToJSON,
+		},
+		{
+			scenario: "json to route",
+			function: testJSONToRoute,
+		},
+		{
+			scenario: "is balancer defined",
+			function: testIsBalancerDefined,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.scenario, func(t *testing.T) {
+			test.function(t)
+		})
+	}
+}
+
+func testNewDefinitions(t *testing.T) {
 	definition := NewDefinition()
 	assert.NotNil(t, definition)
 }
 
-func TestSuccessfulValidation(t *testing.T) {
+func testSuccessfulValidation(t *testing.T) {
 	definition := Definition{
 		ListenPath:  "/*",
 		UpstreamURL: "http://test.com",
@@ -22,7 +66,7 @@ func TestSuccessfulValidation(t *testing.T) {
 	assert.True(t, isValid)
 }
 
-func TestEmptyListenPathValidation(t *testing.T) {
+func testEmptyListenPathValidation(t *testing.T) {
 	definition := Definition{}
 	isValid, err := definition.Validate()
 
@@ -30,7 +74,7 @@ func TestEmptyListenPathValidation(t *testing.T) {
 	assert.False(t, isValid)
 }
 
-func TestInvalidTargetURLValidation(t *testing.T) {
+func testInvalidTargetURLValidation(t *testing.T) {
 	definition := Definition{
 		ListenPath:  " ",
 		UpstreamURL: "wrong",
@@ -41,27 +85,70 @@ func TestInvalidTargetURLValidation(t *testing.T) {
 	assert.False(t, isValid)
 }
 
-func TestRouteToJSON(t *testing.T) {
+func testRouteToJSON(t *testing.T) {
+	expectedJSON := `
+	{
+		"proxy":{
+			"insecure_skip_verify":false,
+			"append_path":false,
+			"enable_load_balancing":false,
+			"methods":[
+
+			],
+			"hosts":[
+
+			],
+			"preserve_host":false,
+			"listen_path":"",
+			"upstream_url":"",
+			"strip_path":false,
+			"upstreams":{
+				"balancing":"",
+				"targets":[
+
+				]
+			}
+		}
+	}
+	`
 	definition := NewDefinition()
 	route := NewRoute(definition)
 	json, err := route.JSONMarshal()
+
 	assert.NoError(t, err)
-	assert.JSONEq(
-		t,
-		`{"proxy": {"insecure_skip_verify": false, "append_path":false, "enable_load_balancing":false, "methods":[], "hosts":[], "preserve_host":false, "listen_path":"", "upstream_url":"", "strip_path":false, "upstreams": {"balancing": "", "targets": [] }}}`,
-		string(json),
-	)
+	assert.JSONEq(t, expectedJSON, string(json))
 }
 
-func TestJSONToRoute(t *testing.T) {
-	route, err := JSONUnmarshalRoute([]byte(`{"proxy": {"insecure_skip_verify": false, "append_path":false, "enable_load_balancing":false, "methods":[], "hosts":[], "preserve_host":false, "listen_path":"", "upstream_url":"/*", "strip_path":false}}`))
+func testJSONToRoute(t *testing.T) {
+	route, err := JSONUnmarshalRoute([]byte(`
+	{
+		"proxy":{
+			"insecure_skip_verify":false,
+			"append_path":false,
+			"enable_load_balancing":false,
+			"methods":[],
+			"hosts":[],
+			"preserve_host":false,
+			"listen_path":"",
+			"upstream_url":"/*",
+			"strip_path":false
+		}
+	}`))
 
 	assert.NoError(t, err)
 	assert.IsType(t, &Route{}, route)
 }
 
-func TestJSONToRouteError(t *testing.T) {
+func testJSONToRouteError(t *testing.T) {
 	_, err := JSONUnmarshalRoute([]byte{})
-
 	assert.Error(t, err)
+}
+
+func testIsBalancerDefined(t *testing.T) {
+	definition := NewDefinition()
+	assert.False(t, definition.IsBalancerDefined())
+
+	target := &Target{Target: "http://localhost:8080/api-name"}
+	definition.Upstreams.Targets = append(definition.Upstreams.Targets, target)
+	assert.True(t, definition.IsBalancerDefined())
 }