+++ pkg/store/in_memory.go
@@ -1,11 +1,23 @@
 package store
 
-import "sync"
+import (
+	"encoding/json"
+	"sync"
+
+	"github.com/hellofresh/janus/pkg/notifier"
+	log "github.com/sirupsen/logrus"
+)
+
+type noopNotification struct {
+	topic string
+	data  []byte
+}
 
 // InMemoryStore is the redis store.
 type InMemoryStore struct {
 	sync.Mutex
-	data map[string]string
+	data   map[string]string
+	client chan noopNotification
 }
 
 // NewInMemoryStore returns an instance of memory store.
@@ -16,7 +28,8 @@ func NewInMemoryStore() *InMemoryStore {
 // NewInMemoryStoreWithOptions returns an instance of memory store with custom options.
 func NewInMemoryStoreWithOptions() *InMemoryStore {
 	return &InMemoryStore{
-		data: make(map[string]string),
+		data:   make(map[string]string),
+		client: make(chan noopNotification),
 	}
 }
 
@@ -50,6 +63,25 @@ func (s *InMemoryStore) Set(key string, value string, expire int64) error {
 	defer s.Unlock()
 
 	return s.set(key, value, expire)
+}
+
+// Publish publishes in memory
+func (s *InMemoryStore) Publish(topic string, data []byte) error {
+	s.client <- noopNotification{topic, data}
+	return nil
+}
+
+// Subscribe subscribes to messages in memory
+func (s *InMemoryStore) Subscribe(channel string, callback func(notifier.Notification)) error {
+	for v := range s.client {
+		notification := notifier.Notification{}
+		if marshallErr := json.Unmarshal(v.data, &notification); marshallErr != nil {
+			log.WithError(marshallErr).Error("Unmarshalling message body failed, malformed")
+			return marshallErr
+		}
+		callback(notification)
+	}
+	return nil
 }
 
 func (s *InMemoryStore) exists(key string) (bool, error) {