+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 3.5.0
+
 ## Added
 - Check GitHub permissions. Sets `is_admin` into the jwt token when the chosen provider is Github
 - Jaeger support as distributed tracing backend